Promote Pi05 stages to generic VLA stages
